% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pop_data}
\alias{get_pop_data}
\title{Get Uk population data}
\usage{
get_pop_data()
}
\value{
a tibble containing UK population data
}
\description{
UK population data for Males, Females and Total

One Year per row
One column for each gender / age combinations
Eg  Total_Age_17, Male_Age_21, Female_Age_30
}
