% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transpose_pop}
\alias{transpose_pop}
\title{Transpose population table from Age in Rows and Year in Columns
to Year in rows and Age in columns}
\usage{
transpose_pop(tbl)
}
\arguments{
\item{tbl}{tibble: must have Age and table_name columns}
}
\value{
tibble: the transposed table
}
\description{
Transpose population table from Age in Rows and Year in Columns
to Year in rows and Age in columns
}
